Code in progress
Attempts to loop through a tier of nodes, and when a node is found to
have child nodes, recursively call the function on the next tier down.
This recursive call is complete when no more child nodes are found.
While navigating the DOM tree, any elements that contain the target
class name are pushed up one tier until the function returns to it’s
initial cycle, returning a complete list.